+++ src/main/java/guru/qa/service/OperationHandler.java
@@ -2,5 +2,5 @@ package guru.qa.service;
 
 public interface OperationHandler {
 
-    int invoke(int first, int second);
+    double invoke(int first, int second);
 }
